Added redirection

See merge request technion-cs/winter-2022/234123-operating-systems!12

+++ hw1/BuiltIn.cpp
@@ -4,13 +4,16 @@
 #include <stdio.h>
 #include <cstring>
 #include <iostream>
+#include <fcntl.h>
 
 #include "BuiltIn.h"
 #include "Shell.h"
 #include "Utility.h"
 #include "JobsList.h"
 
-set<string> BuiltinTable{"cd", "chprompt", "showpid", "pwd", "jobs", "kill", "fg", "bg", "quit"};
+#define MAX_SIZE_TEXT 4096
+
+set<string> BuiltinTable{"cd", "chprompt", "showpid", "pwd", "jobs", "kill", "fg", "bg", "quit","head"};
 
 void ShowPidCommand::execute()
 {
@@ -100,7 +103,13 @@ void KillCommand::execute()
         return;
     }
     // We can now send the signal to the process.
-    int result = kill(job->p_id, stoi(signal_num));
+    int result;
+    if ( signal == SIGCONT) {
+        result = job->_continue_();
+    }
+    else{
+        result = kill(job->p_id, stoi(signal_num));
+    }
     if (result != 0)
     {
         // We have an error.
@@ -256,3 +265,71 @@ void QuitCommand::execute()
     // We can exit the shell completely.
     exit(0);
 }
+
+void HeadCommand::execute() {
+
+    SmallShell &smash = SmallShell::getInstance();
+    vector<string> args = smash.curr_arguments;
+    size_t num_of_lines = 10;
+
+    if (args[1].empty()){
+        cout << "smash error: head: not enough arguments"<<endl;
+        return;
+    }
+
+    if (isNumber(args[1])) num_of_lines = stoi(args[1]);
+
+    int inFile = open(args[2].c_str(),O_RDONLY);
+    if ( inFile == -1){
+        perror("“smash error: open failed");
+        return;
+    }
+
+    char c , buffer[MAX_SIZE_TEXT];
+    size_t index = 0, num_of_written_lines = 1;
+    ssize_t r_result, w_result;
+
+    while ((r_result = read(inFile, &c, 1)) != 0) {
+        if (r_result <= 0) {
+            if ( r_result == 0) {
+                break;
+            }
+            perror("smash error: read failed");
+            break;
+        }
+
+        // Check if the current character is a new line (the line ends here)
+        if (c == '\n') {
+            buffer[index] = c;
+            buffer[index + 1] = '\0';
+            c = 0;
+            index = 0;
+
+            // Print the line
+            w_result = 0;
+            ssize_t buffer_length = strlen(buffer);
+            while (w_result != buffer_length) {
+                ssize_t res = write(0, buffer + w_result, buffer_length - w_result);
+
+                if (w_result < 0) {
+                    perror("smash error: write failed");
+                    break;
+
+                }
+
+                w_result += res;
+            }
+
+            // Stop if we read 10 lines already
+            if (num_of_written_lines == num_of_lines) {
+                break;
+            }
+
+            num_of_written_lines++;
+        } else {
+            buffer[index++] = c;
+        }
+    }
+        close(inFile);
+    }
+
